Tweak single threaded reference presentation
We have only one data point for BLAKE3 and SHA256, but we want to show
this as a line in the graph instead of single data point. Fix by adding
a fake run with the maximum number of threads. There may be a better way
to do this, but this works.