Replace the semaphore with atomics
sem_* functions are deprecated on macOS.

Using atomic-fench synchronization. The thread calling
submission_complete() use atomic store to set sub->completed. The thread
checking if the submission was completed use atomic load to get the
value, and atomic fence to synchronize with the atomic store in the
complete tread.

When waiting for in-flight submissions the main thread use polling. This
may consume entire cpu while waiting, but it happens only when calling
blkhash_final(). If the user wait for all completions, the polling is
expected to be very short.

Tested on x86_64, needs testing on Arm.

[1] https://en.cppreference.com/w/cpp/atomic/atomic_thread_fence#Atomic-fence_synchronization