Add explicit label instead of using filename
Using the file name for the legend repeats too much information that can
be included in the title. Add a label argument to bench.blksum(),
bench.mmap(), and bench.b3sum() and use it in the benchmark scripts.

Update the existing results and rebuild the graphs using the new label.